libpud,libwar2: implement pseudo-exceptions for error handling
libpud & cie were cursed with the READx macros, that were here to do
easy error handling. Unfortunately, they were disgusting and relied on
GNU C extensions. For VS compatibility, these macros are gone, and
replaced by a common framework of pseudo-exceptions.

We now define a "catch" block prior doing reads, which will be executed
upon failure.

It still is not that great, but at least this should be portable.
This also allows for a bit of refactoring.