libpud: swap icons death_and_decay and death_coil
I've written icon numbers using as a reference an old game version. I'm
pretty sure the first Mac version of Tides of Darkness used this. Maybe
the DOS version too.

Anyway, let's swap the icons so we have an API that is consistent with
Battle Net Edition, which clearly is the most widespread.

Yes, this is an API break.

Fixes #6